backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "pypdf>=5.4.0",
    "ragas>=0.2.15",
    "sentence-transformers>=4.1.0",
    "python-dotenv>=0.21.0", # Added for .env file loading
    "sacrebleu>=2.5.1",
    "rouge-score>=0.1.2",
    "rapidfuzz>=3.13.0",
    "langchain-chroma>=0.2.3",
    "langchain-huggingface>=0.2.0",
]
